step3 refactor: 즐겨찾기 도메인 이동
