--- src/main/java/nextstep/subway/applicaion/FavoriteQueryService.java
@@ -1,9 +1,0 @@
-package nextstep.subway.applicaion;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional(readOnly = true)
-public class FavoriteQueryService {
-}
