+++ src/main/java/nextstep/subway/applicaion/dto/FavoriteRequest.java
@@ -1,9 +1,9 @@
-package nextstep.member.application.dto;
+package nextstep.subway.applicaion.dto;
 
 public class FavoriteRequest {
-  private Long source;
+  private final Long source;
 
-  private Long target;
+  private final Long target;
 
   public FavoriteRequest(Long source, Long target) {
     this.source = source;
