+++ src/main/java/nextstep/subway/domain/FavoriteRepository.java
@@ -1,4 +1,4 @@
-package nextstep.member.domain;
+package nextstep.subway.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
